Very basic and fast impl.

+++ common/src/main/java/com/pinkydev/common/model/User.kt
@@ -3,5 +3,5 @@ package com.pinkydev.common.model
 data class User(
     val id: Int = 0,
     val name: String = "",
-    val balance: Float = 100.0f
+    val balance: Int = 100
 )
